# docs/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Plataforma de Apostas - API Core",
    "description": "API Gateways para Wallet, Bets, KYC, Reporting e Responsible Gaming. Base para a implementação e certificação junto ao SPA/MF.",
    "version": "1.0.0",
    "contact": { "name": "Tech", "email": "tech@platform.com.br" }
  },
  "servers": [
    { "url": "https://api.platform.com.br/v1", "description": "Produção" },
    { "url": "https://api-uat.platform.com.br/v1", "description": "Homologação (certificação)" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "BetStatus": {
        "type": "object",
        "properties": {
          "bet_id": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["ACCEPTED", "SETTLED_WIN", "SETTLED_LOSS", "CASHED_OUT", "CANCELLED"] },
          "payout": { "type": "number", "format": "double" }
        }
      },
...
      "SelfExclusion": {
        "type": "object",
        "required": ["duration"],
        "properties": {
          "duration": { "type": "string", "enum": ["24H", "7D", "30D", "90D", "6M", "PERMANENT"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/wallet/webhooks/pix": {
      "post": {
        "tags": ["Wallet"],
        "summary": "Webhook de confirmação PIX do PSP",
        "operationId": "pixWebhook",
        "security": [{ "apiKey": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["event", "payload", "signature"], "properties": {
          "event": { "type": "string", "example": "pix.payment.confirmed" },
          "payload": { "type": "object" },
          "signature": { "type": "string", "description": "Assinatura HMAC do PSP" }
        } } } } },
        "responses": { "200": { "description": "Recebido" } }
      }
    },
...
    "/bets/{betId}/cashout": {
      "post": {
        "tags": ["Bets"],
        "summary": "Cashout de aposta aberta",
        "operationId": "cashout",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "betId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Cashout executado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BetStatus" } } } } }
      }
    },
...
    "/rg/self-exclusion": {
      "post": {
        "tags": ["ResponsibleGaming"],
        "summary": "Autoexclusão",
        "operationId": "selfExclude",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SelfExclusion" } } } },
        "responses": { "200": { "description": "Bloqueio aplicado", "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string", "example": "EXCLUDED" }, "until": { "type": "string", "format": "date-time" } } } } } } }
      }
    },
    "/rg/reality-check": {
      "post": {
        "tags": ["ResponsibleGaming"],
        "summary": "Confirmar realidade check (60 min)",
        "operationId": "realityCheck",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Sessão atualizada", "content": { "application/json": { "schema": { "type": "object", "properties": { "acknowledged": { "type": "boolean" }, "session_seconds": { "type": "integer" }, "net_result": { "type": "number" } } } } } } }
      }
    },
...
  }
}